+++ accounting/state/limit_state.go
@@ -7,8 +7,17 @@ import (
 
 type (
 	LimitState interface {
+		Import(Snapshot)
+		Export() Snapshot
 		check() (bool, int, time.Time)
 		update(int, time.Time)
+	}
+
+	Snapshot struct {
+		Timeout  time.Duration
+		Datetime []time.Time
+		Pattern  string
+		Pointer  int
 	}
 
 	updateRequest struct {