+++ log_code/code.go
@@ -1,18 +1,23 @@
 package log_code
 
 const (
-	WarnProxyGrpcHandler                = 604 //metadata: {"typeData":"", "method":""}
-	WarnConvertErrorDataMarshalResponse = 605
-	WarnJournalCouldNotWriteToFile      = 607
-
 	WarnHttpServerShutdown = 602
 	ErrorHttpServerListen  = 603
 
+	WarnProxyGrpcHandler                = 604 //metadata: {"typeData":"", "method":""}
+	WarnConvertErrorDataMarshalResponse = 605
+
 	ErrorClientGrpc = 606
-	ErrorClientHttp = 611
+
+	WarnJournalCouldNotWriteToFile = 607
 
 	ErrorClientJournal = 608
 	ErrorClientRedis   = 609
+	ErrorClientHttp    = 611
 
 	ErrorAuthenticate = 612
+
+	ErrorClientDatabase     = 615
+	ErrorSnapshotAccounting = 616
+	ErrorUnloadAccounting   = 617
 )